+++ .github/workflows/build.yml
@@ -434,14 +434,13 @@ jobs:
           python --version
           python -m pip install -U pytest-benchmark
           rm -rf tensorflow_io
-          (cd tests && python -m pytest -s -v test_lmdb_eager.py)
-          (python -m pytest -s -v test_image_eager.py -k "webp or ppm or bmp or bounding or exif or hdr or openexr or tiff or avif")
-          (python -m pytest -s -v test_serialization_eager.py)
-          (python -m pytest -s -v test_io_dataset_eager.py -k "numpy or hdf5 or audio or to_file")
-          (python -m pytest -s -v test_http_eager.py)
+          (cd tests && python -m pytest -s -v test_lmdb.py)
+          (python -m pytest -s -v test_image.py -k "webp or ppm or bmp or bounding or exif or hdr or openexr or tiff or avif")
+          (python -m pytest -s -v test_serialization.py)
+          (python -m pytest -s -v test_io_dataset.py -k "numpy or hdf5 or audio or to_file")
+          (python -m pytest -s -v test_http.py)
           python -m pip install google-cloud-bigquery-storage==0.7.0 google-cloud-bigquery==1.22.0 fastavro
-          (python -m pytest -s -v test_bigquery_eager.py)
-          (python -m pytest -s -v test_dicom_eager.py)
+          (python -m pytest -s -v test_bigquery.py)
           (python -m pytest -s -v test_dicom.py)
 
   release:
