+++ tests/test_audio_ops.py
