+++ tests/test_genome.py
@@ -19,8 +19,6 @@ import os
 import numpy as np
 
 import tensorflow as tf
-
-tf.compat.v1.disable_eager_execution()
 import tensorflow_io as tfio  # pylint: disable=wrong-import-position
 
 fastq_path = os.path.join(
@@ -30,13 +28,8 @@ fastq_path = os.path.join(
 
 def test_genome_fastq_reader():
     """test_genome_fastq_reader"""
-    g1 = tf.compat.v1.Graph()
 
-    with g1.as_default():
-        data = tfio.genome.read_fastq(filename=fastq_path)
-
-    sess = tf.compat.v1.Session(graph=g1)
-    data_np = sess.run(data)
+    data = tfio.genome.read_fastq(filename=fastq_path)
 
     data_expected = [
         b"GATTACA",
@@ -52,8 +45,8 @@ def test_genome_fastq_reader():
         b"FAD",
     ]
 
-    assert np.all(data_np.sequences == data_expected)
-    assert np.all(data_np.raw_quality == quality_expected)
+    assert np.all(data.sequences == data_expected)
+    assert np.all(data.raw_quality == quality_expected)
 
 
 def test_genome_sequences_to_onehot():
@@ -189,12 +182,10 @@ def test_genome_sequences_to_onehot():
         [[0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 1, 0]],
     ]
 
-    with tf.compat.v1.Session() as sess:
-        raw_data = tfio.genome.read_fastq(filename=fastq_path)
-        data = tfio.genome.sequences_to_onehot(sequences=raw_data.sequences)
-        out = sess.run(data)
+    raw_data = tfio.genome.read_fastq(filename=fastq_path)
+    data = tfio.genome.sequences_to_onehot(sequences=raw_data.sequences)
 
-    assert np.all(out.to_list() == expected)
+    assert np.all(data.to_list() == expected)
 
 
 def test_genome_phred_sequences_to_probability():
@@ -210,27 +201,20 @@ def test_genome_phred_sequences_to_probability():
         0.00019952621369156986,
     ]
 
-    with tf.compat.v1.Session() as sess:
-        example_quality = tf.constant(example_quality_list)
-        converted_phred = tfio.genome.phred_sequences_to_probability(example_quality)
-        out = sess.run(converted_phred)
+    example_quality = tf.constant(example_quality_list)
+    converted_phred = tfio.genome.phred_sequences_to_probability(example_quality)
 
     # Compare flat values
-    assert np.allclose(out.flat_values.flatten(), expected_probabilities)
+    assert np.allclose(
+        converted_phred.flat_values.numpy().flatten(), expected_probabilities
+    )
     # Ensure nested array lengths are correct
     assert np.all(
-        [len(a) == len(b) for a, b in zip(out.to_list(), example_quality_list)]
+        [
+            len(a) == len(b)
+            for a, b in zip(converted_phred.to_list(), example_quality_list)
+        ]
     )
-
-
-def test_genome_phred_sequences_to_probability_with_other_genome_ops():
-    """Test quality op in graph with read_fastq op, ensure no errors"""
-    with tf.compat.v1.Session() as sess:
-        raw_data = tfio.genome.read_fastq(filename=fastq_path)
-        data = tfio.genome.phred_sequences_to_probability(
-            phred_qualities=raw_data.raw_quality
-        )
-        sess.run(data)
 
 
 if __name__ == "__main__":
