+++ tests/test_image.py
