+++ tests/test_text.py
